requirements-minimal.txt
# Core dependencies - Actually used in the project
streamlit==1.52.1
python-dotenv==1.2.1

# Machine Learning / Deep Learning
tensorflow==2.20.0
numpy==2.2.6
scikit-learn==1.7.2
pillow==12.0.0

# LangChain ecosystem
langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-google-genai==4.1.1

# Vector database
chromadb==1.3.7

# Embeddings
sentence-transformers==5.2.0
huggingface-hub==0.36.0

# Utilities
tqdm==4.67.1
